Changes to some columns to use enum

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->integer('age');
+            $table->unsignedTinyInteger('age');
             $table->string('location');
             $table->char('phone_number', 10);
             $table->rememberToken();
